--- jdbc/src/main/java/nextstep/jdbc/ExtractStrategy.java
@@ -1,11 +1,0 @@
-package nextstep.jdbc;
-
-import java.sql.ResultSet;
-import java.sql.SQLException;
-import java.util.List;
-
-@FunctionalInterface
-interface ExtractStrategy<T> {
-
-    List<T> extract(RowMapper<T> rowMapper, ResultSet resultSet) throws SQLException;
-}
