+++ jdbc/src/main/java/nextstep/jdbc/JdbcTemplate.java
@@ -2,7 +2,9 @@ package nextstep.jdbc;
 
 import java.sql.Connection;
 import java.sql.PreparedStatement;
+import java.sql.ResultSet;
 import java.sql.SQLException;
+import java.util.ArrayList;
 import java.util.List;
 import javax.sql.DataSource;
 import org.springframework.jdbc.datasource.DataSourceUtils;
@@ -22,7 +24,7 @@ public class JdbcTemplate {
     }
 
     public <T> T queryForObject(final String sql, final RowMapper<T> rowMapper, final Object... params) {
-        return query(sql, statement -> QueryExecutor.executeQuery(rowMapper, statement), params);
+        return query(sql, statement -> QueryExecutor.executeQueryForObject(rowMapper, statement), params);
     }
 
     public <T> List<T> queryForList(final String sql, final RowMapper<T> rowMapper, final Object... params) {
@@ -48,5 +50,43 @@ public class JdbcTemplate {
 
     public DataSource getDataSource() {
         return dataSource;
+    }
+
+    private static class QueryExecutor {
+        private QueryExecutor() {
+        }
+
+        private static <T> T executeQueryForObject(final RowMapper<T> rowMapper, final PreparedStatement statement) {
+            return getSingleResult(executeQuery(rowMapper, statement));
+        }
+
+        private static <T> List<T> executeQueryForList(final RowMapper<T> rowMapper,
+                                                       final PreparedStatement statement) {
+            return executeQuery(rowMapper, statement);
+        }
+
+        private static <T> List<T> executeQuery(final RowMapper<T> rowMapper, final PreparedStatement statement) {
+            try (final ResultSet resultSet = statement.executeQuery()) {
+                final List<T> result = new ArrayList<>();
+                while (resultSet.next()) {
+                    result.add(rowMapper.map(resultSet));
+                }
+                return result;
+            } catch (final SQLException e) {
+                throw new DataAccessException("query exception!", e);
+            }
+        }
+
+        private static <T> T getSingleResult(final List<T> results) {
+            if (results.size() > 1) {
+                throw new DataAccessException("more than one result!");
+            }
+
+            if (results.isEmpty()) {
+                throw new DataAccessException("query result is null");
+            }
+
+            return results.get(0);
+        }
     }
 }
